refactor(formatter): split streaming helper modules
